Move inline tests to device_section.rs and priority_section.rs for direct test coverage
